Added timeout handling to the power_board_panel to address ticket #1065. The panel will timeout if it does not get a power board diagnostic for 5 seconds of real time or 5 seconds of ros time, whichever comes last. When it is disabled, the panel disables all its buttons and text fields.
Modified the power_node so that it only publishes diagnostics when it has
actually received a message from the power board. This way if the power
board gets disconnected, the GUI will notice. (The drawback of this
approach is that in the rare case when we can send messages to the power
board, but cannot receive messages from it, the GUI will still prevent
messages from being sent. Probably irrelevant.) 

Added a power_node_simulator to simulate the power board during testing.



git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@13875 f5854215-dd47-0410-b2c4-cdd35faa7885

Former-commit-id: f8878c0da5a5c8a102e9e877c303a73e528a4252