Added triggered mode to the forearm camera driver, and added a timestamp to the outgoing image. The timestamp should be very accurate (1 ms) in the triggered case, and is usually equally accurate in the untriggered case.
So far the camera is occasionally triggering early, a condition which is
not currently well handled, so the timestamp will be incorrect when this
happens.

I had to modify the interface to the camera library slightly in order to
get as accurate timing information as posible. While I was at it a also
removed a few annoying warnings.



git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@13989 f5854215-dd47-0410-b2c4-cdd35faa7885

Former-commit-id: 174826922ad1244622de4c51168b124709711d7a