The comment is also for 16232.
I added a node snapper.py that gets images and submits them directly to MTurk.
I added a node that re-sends images on button ("m"). The image then goes to snapper.

I added session_2_messages.py that gets all annotations from the server and sends them as messages in a fixed coordinate system. The timestamps in the header are 1ns offset from the original stereo frame.

In this revision I deleted two junk files that slipped into the r16232.



git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@16236 f5854215-dd47-0410-b2c4-cdd35faa7885

Former-commit-id: 1a51f1e4305fe33ebddb8aa1536e296696225296