Planar patch map takes planar patches computed by planar_patch_map::planar_patch_map_node, projects them into the image and attaches the labels of 2D polygons to 3D structures that project inside the annotation polygon.
The annotated patches are collected by the map assembler (based on the point cloud assembler) and stored in memory.

QueryAnnotatedMap service allows to get only the map with the specified annotations in a certain time range.

A snapshotter node(annotated_map_query_snapshotter) periodically queries the service and sends the messages with the annotated map out.





git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@16245 f5854215-dd47-0410-b2c4-cdd35faa7885

Former-commit-id: 3ea61055dfa8694871179474ae53719301bbe015